+++ R/topic_modeling_utilities.R
@@ -3,7 +3,7 @@
 #' @param model A list (or S3 object) with three named matrices: phi, theta, and gamma.
 #'        These conform to outputs of many of \link[textmineR]{textmineR}'s native
 #'        topic modeling functions such as \link[textmineR]{FitLdaModel}. 
-#' @return An object of class \code{data.frame} with 6 columns: 'topic' is the 
+#' @return An object of class \code{data.frame} or \code{tibble} with 6 columns: 'topic' is the 
 #'         name of the topic, 'prevalence' is the rough prevalence of the topic 
 #'         in all documents across the corpus, 'coherence' is the probabilistic
 #'         coherence of the topic, 'top_terms_phi' are the top 5 terms for each
@@ -80,7 +80,11 @@ SummarizeTopics <- function(model){
                     top_terms_phi = tt_phi,
                     top_terms_gamma = tt_gamma,
                     stringsAsFactors = FALSE)
-  
+
+    if ("tibble" %in% row.names(utils::installed.packages())) {
+        out <- tibble::as_tibble(out)
+    }
+ 
   out
 }
 
@@ -215,7 +219,11 @@ LabelTopics <- function(assignments, dtm, M=2){
 #' @description Takes topics by terms matrix and returns top M terms for each topic
 #' @param phi A matrix whose rows index topics and columns index words
 #' @param M An integer for the number of terms to return
-#' @return Returns a \code{data.frame} whose columns correspond to a topic and
+#' @param return_matrix Do you want a \code{matrix} or \code{data.frame}/\code{tibble}
+#'   returned? Defaults to \code{TRUE}.
+#' @return 
+#' If \code{return_matrix = TRUE} (the default) then a matrix. Otherwise,
+#' returns a \code{data.frame} or \code{tibble} whose columns correspond to a topic and
 #' whose m-th row correspond to the m-th top term from the input \code{phi}.
 #' @export
 #' @examples
@@ -226,11 +234,20 @@ LabelTopics <- function(assignments, dtm, M=2){
 #' 
 #' str(top_terms)
 
-GetTopTerms <- function(phi, M){
+GetTopTerms <- function(phi, M, return_matrix = TRUE){
   
   result <- apply(phi, 1, function(x){
     names(x)[ order(x, decreasing=TRUE) ][ 1:M ]
   })
+  
+  if (! return_matrix) {
+    if ("tibble" %in% row.names(utils::installed.packages())) {
+      result <- tibble::as_tibble(result)
+    } else {
+      result <- as.data.frame(result)
+    }
+  }
+
   
   return(result)
 }