+++ NEWS.md
@@ -1,9 +1,39 @@
+# textmineR 3.0.6
+This version is a patch. In this version I have
+
+* Updated the URL to a working paper cited in the topic modeling vignette.
+
+# textmineR 3.0.5
+This version is a patch. In this version I have
+
+* Fixed a bug in `CalcHellignerDist()` and `CalcJSDivergence()` that sometimes
+  caused inputs to be overwritten.
+* Fixed some typos in the vignette for topic modeling
+* Updated the documentation on `FitCtmModel()` to better explain how to pass
+  control arguments to CTM's underlying function.
+* Enabled return of a `tibble` or `data.frame` (instead of only `data.frame`) in
+  the following functions: `SummarizeTopics`, `GetTopTerms`, `TermDocFreq`
+  (Thanks to Mattias for the PR)
+  
+# textmineR 3.0.4
+This version is a patch. In this version I have
+
+* Removed unconditional stripping in MAKEVARs as specified by CRAN
+* Improved outputs of `FitLdaModel`
+
+# textmineR 3.0.3
+This version is a patch. In this version I have
+
+* fixed an error related to the `update.lda_topic_model` method.
+* added a method `posterior.lda_topic_model` to sample from the posterior of an
+  LDA topic model. 
+
 # textmineR 3.0.2
 This version is a patch. In this version I have
 
 * changed some elements of NAMESPACE to pass additional CRAN checks.
 * added an update method for the lda_topic_model class. This allows users to add
-  documents to an existing model (and even add new topics) without chaning the 
+  documents to an existing model (and even add new topics) without changing the 
   indices of previously-trained topics. e.g. topic 5 is still topic 5.
 * added a vignette for using `tidytext` alongside `textmineR`
 
@@ -96,7 +126,7 @@ This version significantly changes textmineR.
 * Changes to how the package interfaces with Rcpp
 
 # textmineR 2.0.5
-* Add `verbose` option to `CreateDtm` and `CreateTcm` to surpress status messages.
+* Add `verbose` option to `CreateDtm` and `CreateTcm` to supress status messages.
 * Add function `GetVocabFromDtm` to get `text2vec` vocabulary object from a `dgCMatrix` 
   document term matrix.
   
@@ -134,7 +164,7 @@ This version significantly changes textmineR.
 # textmineR 2.0.0
 
 * Vec2Dtm is now deprecated in favor of CreateDtm
-* A function, CreateTcm, now exists to create term co-occurence matrices
+* A function, CreateTcm, now exists to create term co-occurrence matrices
 * CreateDtm and CreateTcm are implemented with a parallel C++ back end through the text2vec library
   - the implementation is _much_ faster! I've clocked 2X - 10X speedups, depending on options
   - adds external dependencies - C++ compiler and GNU make - and takes away an external
@@ -146,7 +176,7 @@ This version significantly changes textmineR.
     custom or non-English stopwords
   - Now the stemming argument allows for passing of stem/lemmatization functions.
 * Function for fitting correlated topic models
-* Function to turn a document term matrix to term co-occurence matrix
+* Function to turn a document term matrix to term co-occurrence matrix
 * Allowed LabelTopics to use unigrams, if you want. (n-grams are still better.)
 * More robust error checking for CalcTopicModelR2 and CalcLikelihood
 * All function arguments use "_", not ".".
@@ -162,7 +192,7 @@ This version significantly changes textmineR.
 * Added modeling capability for latent semantic analysis in FitLsaModel()
 * Added CalcProbCoherence() function which replaces ProbCoherence() and can calculate
   probabilistic coherence for the whole phi matrix.
-* Added data from NIH research grants instead of borrowd data from tm
+* Added data from NIH research grants instead of borrowed data from tm
 * Removed qcq data 
 * Added variational em method for FitLdaModel()
 * Added function to represent document clustering as a topic model Cluster2TopicModel()